CartItemTest passing
The expiration for CartItem was not working correctly.

The CartItem has its expiration time calculated based on the time of its creation + the modifier in the unity specified by the mode (seconds, in the case of the test), at the constructor.

The method is_available should return false while the current time is less than the expiration time, but true as soon the expiration time is reached. Therefore, the expression operator in is_available should be changed from < to <=, to include the expiration time itself as a time from which the item becomes available.